COMMERCE-2820 - refactor, move JsonReader to util package
